Refactor to use hooks
