+++ packages/block-library/src/editor.scss
@@ -19,6 +19,7 @@
 @import "./list/editor.scss";
 @import "./more/editor.scss";
 @import "./navigation-menu/editor.scss";
+@import "./navigation-menu-item/editor.scss";
 @import "./nextpage/editor.scss";
 @import "./paragraph/editor.scss";
 @import "./preformatted/editor.scss";
