+++ packages/block-library/src/navigation-menu/menu-item-inserter.js
@@ -10,69 +10,53 @@ import {
 	NavigableMenu,
 	TextControl,
 } from '@wordpress/components';
-import { Component } from '@wordpress/element';
+import { useState, useMemo, useCallback } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { createBlock } from '@wordpress/blocks';
 import { isURL } from '@wordpress/url';
 
-class MenuItemInserter extends Component {
-	constructor() {
-		super( ...arguments );
-		this.state = {
-			currentSearchInput: '',
-			isUrlInput: false,
-		};
-		this.onChangeSearchInput = this.onChangeSearchInput.bind( this );
-	}
-
-	onChangeSearchInput( value ) {
-		this.setState( {
-			isUrlInput: isURL( value ),
-			currentSearchInput: value,
+function MenuItemInserter( { insertMenuItem } ) {
+	const [ searchInput, setSearchInput ] = useState( '' );
+	const isUrlInput = useMemo( () => isURL( searchInput ), [ searchInput ] );
+	const onMenuItemClick = useCallback( () => {
+		insertMenuItem( {
+			destination: searchInput,
 		} );
-	}
+	}, [ insertMenuItem, searchInput ] );
 
-	render() {
-		const { insertMenuItem } = this.props;
-		const { currentSearchInput, isUrlInput } = this.state;
-		return (
-			<Dropdown
-				className="wp-block-navigation-menu__inserter"
-				position="bottom center"
-				renderToggle={ ( { isOpen, onToggle } ) => (
-					<IconButton
-						icon="insert"
-						label={ __( 'Insert a new menu item' ) }
-						onClick={ onToggle }
-						aria-expanded={ isOpen }
+	return (
+		<Dropdown
+			className="wp-block-navigation-menu__inserter"
+			position="bottom center"
+			renderToggle={ ( { isOpen, onToggle } ) => (
+				<IconButton
+					icon="insert"
+					label={ __( 'Insert a new menu item' ) }
+					onClick={ onToggle }
+					aria-expanded={ isOpen }
+				/>
+			) }
+			renderContent={ () => (
+				<div className="wp-block-navigation-menu__inserter-content">
+					<TextControl
+						value={ searchInput }
+						label={ __( 'Search or paste a link' ) }
+						onChange={ setSearchInput }
 					/>
-				) }
-				renderContent={ () => (
-					<div className="wp-block-navigation-menu__inserter-content">
-						<TextControl
-							value={ currentSearchInput }
-							label={ __( 'Search or paste a link' ) }
-							onChange={ this.onChangeSearchInput }
-						/>
-						{ isUrlInput && (
-							<NavigableMenu>
-								<MenuItem
-									onClick={ () => {
-										insertMenuItem( {
-											destination: currentSearchInput,
-										} );
-									} }
-									icon="admin-links"
-								>
-									{ currentSearchInput }
-								</MenuItem>
-							</NavigableMenu>
-						) }
-					</div>
-				) }
-			/>
-		);
-	}
+					{ isUrlInput && (
+						<NavigableMenu>
+							<MenuItem
+								onClick={ onMenuItemClick }
+								icon="admin-links"
+							>
+								{ searchInput }
+							</MenuItem>
+						</NavigableMenu>
+					) }
+				</div>
+			) }
+		/>
+	);
 }
 
 export default compose( [
